Update dynamic shortcuts whenever SAF roots are changed
Previously, shortcuts didn't get updated when importing an rclone
config. Let's just merge the SAF roots notification and dynamic shortcut
update operations so they're always in sync.
